fix re-render issue with volunteer event page view
