--- src/screens/Events/Admin/index.jsx
@@ -1,3 +1,0 @@
-import EventManager from "./EventManager";
-
-export default EventManager;
