--- src/screens/Events/User/EventManager.jsx
@@ -1,258 +1,0 @@
-import { useSession } from "next-auth/react";
-import { useRouter } from "next/router";
-import React, { useEffect, useState } from "react";
-import Calendar from "react-calendar";
-import "react-calendar/dist/Calendar.css";
-import { Button } from "reactstrap";
-import styled from "styled-components";
-import { fetchEvents } from "../../../actions/queries";
-import variables from "../../../design-tokens/_variables.module.scss";
-import { registerForEvent, updateEvent } from "./eventHelpers";
-import EventTable from "./EventTable";
-
-const Styled = {
-  Container: styled.div`
-    width: 100%;
-    height: 100%;
-    background: ${(props) => props.theme.grey9};
-    padding-top: 1rem;
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-    overflow: hidden;
-  `,
-  Button: styled(Button)`
-    background: ${variables.primary};
-    border: none;
-    color: white;
-    width: 7.5rem;
-    height: 3rem;
-    margin-top: 2rem;
-    margin-bottom: 2vw;
-
-    &:focus {
-      background: white;
-      outline: none;
-      border: none;
-    }
-  `,
-  Content: styled.div`
-    margin-top: 7rem;
-  `,
-  EventContainer: styled.div`
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 1rem;
-  `,
-  Events: styled.div`
-    text-align: left;
-    font-size: 36px;
-    font-weight: bold;
-  `,
-  Date: styled.div`
-    text-align: left;
-    font-size: 28px;
-    font-weight: bold;
-  `,
-  DateRow: styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  `,
-  Back: styled.p`
-    font-size: 14px;
-    margin-left: 10px;
-    padding-top: 8px;
-    text-decoration: underline;
-    color: ${variables.primary};
-    cursor: pointer;
-  `,
-  Left: styled.div`
-    margin-left: 10vw;
-    display: flex;
-    flex-direction: column;
-  `,
-  Right: styled.div`
-    display: flex;
-    flex-direction: column;
-    margin-left: 3vw;
-  `,
-};
-
-const EventManager = ({ user }) => {
-  const [loading, setLoading] = useState(true);
-  const [events, setEvents] = useState([]);
-  const [showRegisterModal, setShowRegisterModal] = useState(false);
-  const [currEvent, setCurrEvent] = useState(null);
-  const [currUser, setCurrUser] = useState(null);
-  const [showBack, setShowBack] = useState(false);
-
-  const router = useRouter();
-
-  if (!user) {
-    const { data: session } = useSession();
-    user = session.user;
-  }
-  useEffect(() => {
-    onRefresh();
-  }, []);
-
-  const [markDates, setDates] = useState([]);
-
-  const onRefresh = () => {
-    setLoading(true);
-    fetchEvents()
-      .then((result) => {
-        if (result && result.data && result.data.events) {
-          // result.data.events = result.data.events.filter(function (event) {
-          //   const currentDate = new Date();
-          //   return new Date(event.date) > currentDate;
-          // });
-          setEvents(result.data.events);
-          setDates(result.data.events);
-        }
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
-
-  const onRegister = async (event) => {
-    const changedEvent = {
-      ...event,
-      volunteers: event.volunteers.concat(user._id),
-    }; // adds userId to event
-    const updatedEvent = await registerForEvent({ user, event: changedEvent }); // updates event in backend
-    setEvents(events.map((e) => (e._id === event._id ? updatedEvent : e))); // set event state to reflect new event
-
-    onRefresh();
-  };
-
-  const onUnregister = async (event) => {
-    const changedEvent = {
-      // remove current user id from event volunteers
-      ...event,
-      volunteers: event.volunteers.filter(
-        (volunteer) => volunteer !== user._id
-      ),
-    };
-    const updatedEvent = await updateEvent(changedEvent);
-    setEvents(events.map((e) => (e._id === event._id ? updatedEvent : e)));
-
-    onRefresh();
-  };
-
-  const [value, setDate] = useState(new Date());
-
-  let splitDate = value.toDateString().split(" ");
-  const [dateString, setDateString] = useState(
-    splitDate[1] + " " + splitDate[2] + " " + splitDate[3]
-  );
-
-  const onChange = (value, event) => {
-    if (Date.now() !== value) setShowBack(true);
-    setDate(value);
-    let datestr = value.toString();
-    let splitDate = value.toDateString().split(" ");
-    let date = splitDate[1] + " " + splitDate[2] + ", " + splitDate[3];
-    setDateString(date);
-    let selectDate = new Date(datestr).toISOString().split("T")[0];
-
-    setLoading(true);
-    fetchEvents(selectDate, selectDate)
-      .then((result) => {
-        if (result && result.data && result.data.events) {
-          setEvents(result.data.events);
-        }
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
-
-  const onRegisterClicked = (event, user) => {
-    setShowRegisterModal(true);
-    setCurrEvent(event);
-    setCurrUser(user);
-    router.replace("/register");
-  };
-
-  const toggleRegisterModal = () => {
-    setShowRegisterModal((prev) => !prev);
-    onRefresh();
-  };
-
-  const formatJsDate = (jsDate, separator = "/") => {
-    return [
-      String(jsDate.getFullYear()).padStart(4, "0"),
-      String(jsDate.getMonth() + 1).padStart(2, "0"),
-      String(jsDate.getDate()).padStart(2, "0"),
-    ].join(separator);
-  };
-
-  const setMarkDates = ({ date, view }, markDates) => {
-    const fDate = formatJsDate(date, "-");
-    let tileClassName = "";
-    let test = [];
-    for (let i = 0; i < markDates.length; i++) {
-      test.push(markDates[i].date.slice(0, 10));
-    }
-    if (test.includes(fDate)) {
-      tileClassName = "marked";
-    }
-    return tileClassName !== "" ? tileClassName : null;
-  };
-
-  const setDateBack = () => {
-    const currentDate = new Date();
-    setDates(currentDate);
-    setDate(currentDate);
-    let splitDate = currentDate.toDateString().split(" ");
-    let date = splitDate[1] + " " + splitDate[2] + ", " + splitDate[3];
-    setDateString(date);
-    setShowBack(false);
-    onRefresh();
-  };
-
-  return (
-    <Styled.Container>
-      <Styled.Left>
-        <Styled.EventContainer>
-          <Styled.Events>Events</Styled.Events>
-          <Styled.DateRow>
-            <Styled.Date>{dateString}</Styled.Date>
-            {showBack && (
-              <Styled.Back onClick={setDateBack}>Back to Today</Styled.Back>
-            )}
-          </Styled.DateRow>
-        </Styled.EventContainer>
-        <Calendar
-          onChange={onChange}
-          value={value}
-          tileClassName={({ date, view }) =>
-            setMarkDates({ date, view }, markDates)
-          }
-        />
-      </Styled.Left>
-      <Styled.Right>
-        <Styled.Content>
-          {events.length == 0 ? (
-            <Styled.Events>No Events Scheduled on This Date</Styled.Events>
-          ) : (
-            <EventTable
-              dateString={dateString}
-              events={events}
-              onRegisterClicked={onRegister}
-              onUnregister={onUnregister}
-              user={user}
-            >
-              {" "}
-            </EventTable>
-          )}
-        </Styled.Content>
-      </Styled.Right>
-    </Styled.Container>
-  );
-};
-
-export default EventManager;
