+++ src/screens/Events/index.jsx
@@ -1,7 +1,5 @@
 import { useSession } from "next-auth/react";
 import React from "react";
-import AdminEventManager from "./Admin";
-import UserEventManager from "./User";
 import EventManager from "./EventManager";
 
 const EventManagerSelector = () => {
@@ -9,11 +7,6 @@ const EventManagerSelector = () => {
     data: { user },
   } = useSession();
 
-  // return user.role === "admin" ? (
-  //   <AdminEventManager user={user} />
-  // ) : (
-  //   <UserEventManager user={user} />
-  // );
   return <EventManager user={user} role={user.role} />
 };
