+++ src/screens/Homepage/index.jsx
@@ -1,7 +1,7 @@
 import { useSession } from "next-auth/react";
 import React from "react";
 import UserEventManager from "./User";
-import EventManager from "../Events/Admin";
+import EventManager from "../Events";
 
 const EventManagerSelector = () => {
   const {
@@ -9,7 +9,7 @@ const EventManagerSelector = () => {
   } = useSession();
 
   if (user.role !== "volunteer") {
-    return <EventManager user={user} />;
+    return <EventManager user={user} role={user.role} />;
   } else {
     return <UserEventManager user={user} />;
   }
